Split AdminProductCard into editing/viewing components; make it possible to see and edit provided ticket types
